requirements.txt
python-telegram-bot[job-queue]==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
googletrans==4.0.0rc1
